edited price on database

+++ bamazon.sql
@@ -9,7 +9,7 @@ CREATE TABLE products (
 
 	 department_name VARCHAR(20) NOT NULL,
      
-     price INTEGER(10) NOT NULL,
+     price DECIMAL(10,2) NOT NULL,
      
      stock_quantity INTEGER(10) NOT NULL,
